refactor APITokens so it's a container component with two paginated children
